packages/python/test/fixtures/33-fastapi-zipped-deps/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
requests
pyyaml

bcrypt
PyJWT

supabase

openai
anthropic
cohere
mistralai
openai-agents
langchain
langchain-community
pinecone

posthog
resend
exa-py

redis
